Wire platform event task launches end to end
