+++ apps/api/src/handlers/slack/events/channel-auto-start-failure.test.ts
@@ -33,6 +33,7 @@ vi.mock('@roomote/env', () => ({
 
 vi.mock('@roomote/cloud-agents/server', () => ({
   ROUTING_AUTO_CONFIRM_TIMEOUT_MS: 0,
+  createFastAgentSlackTaskLauncher: vi.fn(() => vi.fn()),
 }));
 
 vi.mock('@roomote/cloud-agents', () => ({
