+++ apps/api/src/handlers/slack/events/fast-agent-task-launcher.ts
@@ -1,6 +1,5 @@
 import {
-  enqueueTask,
-  getTaskUrl,
+  createFastAgentSlackTaskLauncher,
   type LaunchFastAgentSlackTask,
 } from '@roomote/cloud-agents/server';
 import { type SlackEvent } from '@roomote/slack';
@@ -8,12 +7,6 @@ import {
   type SlackInstallation,
   type SlackUserMapping,
 } from '@roomote/db/server';
-import {
-  ALL_REPOSITORIES,
-  TaskPayloadKind,
-  type StandardTask,
-} from '@roomote/types';
-
 export function createFastAgentTaskLauncher(params: {
   event: SlackEvent;
   slackInstallation: SlackInstallation;
@@ -21,64 +14,14 @@ export function createFastAgentTaskLauncher(params: {
   userId: string;
   teamId: string;
 }): LaunchFastAgentSlackTask {
-  return async ({ prompt, environmentId, parentSessionId, postKickoff }) => {
-    const threadId = params.event.thread_ts || params.event.ts;
-    const task: StandardTask = {
-      type: TaskPayloadKind.StandardTask,
-      payload: {
-        repo: ALL_REPOSITORIES,
-        description: prompt,
-        communicationProvider: 'slack',
-        communicationTeamId: params.teamId,
-        communicationTeamDomain:
-          params.slackInstallation.teamDomain ?? undefined,
-        communicationChannelId: params.event.channel,
-        communicationThreadId: threadId,
-        communicationMessageId: params.event.ts,
-        communicationContextInherited: true,
-        fastAgentSessionId: parentSessionId,
-        fastAgentParent: {
-          sessionId: parentSessionId,
-          slackTeamId: params.teamId,
-          slackChannel: params.event.channel,
-          slackThreadTs: threadId,
-        },
-        ...(environmentId && environmentId !== ALL_REPOSITORIES
-          ? { environmentId }
-          : {}),
-      },
-    };
-    let taskUrl: string | undefined;
-    const launch = await enqueueTask(
-      {
-        task,
-        initiator: { kind: 'user', userId: params.userId },
-        workflow: 'standard',
-        surface: 'slack',
-        trigger: 'message',
-      },
-      {
-        beforeEnqueue: async (taskRun) => {
-          taskUrl = getTaskUrl({
-            taskId: taskRun.taskId,
-            utm: { source: 'slack', campaign: 'fast-delegation' },
-          });
-          await postKickoff({ taskId: taskRun.taskId, taskUrl });
-        },
-      },
-    );
-
-    if (!launch.taskId) {
-      return {
-        success: false,
-        error: 'The task launch did not return a task ID.',
-      };
-    }
-
-    return {
-      success: true,
-      taskId: launch.taskId,
-      taskUrl,
-    };
-  };
+  return createFastAgentSlackTaskLauncher({
+    userId: params.userId,
+    teamId: params.teamId,
+    ...(params.slackInstallation.teamDomain
+      ? { teamDomain: params.slackInstallation.teamDomain }
+      : {}),
+    channelId: params.event.channel,
+    threadTs: params.event.thread_ts || params.event.ts,
+    messageId: params.event.ts,
+  });
 }
