+++ packages/cloud-agents/src/server/fast-agent/index.ts
@@ -3,5 +3,6 @@ export * from './fast-agent-prompt';
 export * from './fast-agent-service';
 export * from './fast-agent-turn-lock';
 export * from './fast-agent-session';
+export * from './fast-agent-task-launcher';
 export * from './fast-agent-tasks';
 export * from './onboarding-task-suggestions-service';
